Add highlight code.
