+++ mikoto/__init__.py
@@ -2,7 +2,7 @@
 
 from mikoto.markdown import render_markdown
 from mikoto.rst import render_rst
-from mikoto.code import render_code
+from mikoto.code import render_code, render_highlight_code
 from mikoto.text import translate_to_unicode
 
 __all__ = ['Mikoto']
@@ -24,3 +24,6 @@ class Mikoto(object):
     @property
     def code(self):
         return render_code(self.unicode)
+
+    def highlight_code(self, path):
+        return render_highlight_code(self.unicode, path)
